added No-Such-Element-Exception to delete-method + test

+++ backend/src/main/java/de/neuefische/backend/service/PositionsService.java
@@ -44,6 +44,11 @@ public class PositionsService {
     }
 
     public void deletePositionById(String id) {
+        if (!positionsRepo.existsById(id)) {
+            throw new NoSuchElementException("Position with Id " + id +  " does not exist.");
+        }
+
         positionsRepo.deleteById(id);
+
     }
 }
